fix codeclimate move methods to ProcessUserData
