+++ app/src/main/java/org/ole/planet/takeout/ProcessUserData.java
@@ -1,16 +1,36 @@
 package org.ole.planet.takeout;
 
 import android.content.SharedPreferences;
+import android.support.design.widget.TextInputLayout;
 import android.support.v7.app.AppCompatActivity;
+import android.view.View;
+import android.view.WindowManager;
+import android.widget.EditText;
 
 import com.google.gson.JsonObject;
 
 import io.realm.Realm;
 
 public abstract class ProcessUserData extends AppCompatActivity {
-
-    public static final String PREFS_NAME = "OLE_PLANET";
     SharedPreferences settings;
+
+    public boolean validateEditText(EditText textField, TextInputLayout textLayout, String err_message) {
+        if (textField.getText().toString().trim().isEmpty()) {
+            textLayout.setError(err_message);
+            requestFocus(textField);
+            return false;
+        } else {
+            textLayout.setErrorEnabled(false);
+        }
+        return true;
+    }
+
+
+    private void requestFocus(View view) {
+        if (view.requestFocus()) {
+            getWindow().setSoftInputMode(WindowManager.LayoutParams.SOFT_INPUT_STATE_ALWAYS_VISIBLE);
+        }
+    }
 
     public void saveUserInfoPref(SharedPreferences settings, String password, realm_UserModel user) {
         this.settings = settings;
@@ -52,5 +72,6 @@ public abstract class ProcessUserData extends AppCompatActivity {
             err.printStackTrace();
         }
     }
+
 
 }
